+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added new Digital Foodsharing Passports for Android and iOS !3591 @SaiCode
 - Food baskets support multiple images now. Image quality issues are resolved. #1797 #1714 #2120 !3304 !3711 @AntonBallmaier
 - Added the option to test a quiz from the quiz editor. #2112 !3612 @AntonBallmaier
+- Ambassadors can now see the previous home district and last pickup of users in their region. #2004 !3675 @AntonBallmaier
 - Added bells for mentions in forum. #1850 !3671 @AntonBallmaier
 - Added a button to the markdown input for mentioning users #1850 !3670 #2153 !3692 @AntonBallmaier
 - Added the option to send a message with a store aplication #2017 !3461 @AntonBallmaier